Working on a script that will redraw an image using just one long squiggle.
The idea is to make a point cloud (with density based on graysale values
of the original image), make a Delaunay triangulation of the point cloud,
then plan the squiggle as a long simple path through the edges of the
triangulation. Then it's straightforward to use quadratic splines to make
the path look like a continuous squiggle. And having the triangulation will
help me place spline control points so that the squiggle never self-intersects.

I am enjoying myself.